requirements.txt
pandas
sqlalchemy
python-dotenv
pyodbc
psycopg2-binary
pyarrow
geopandas
requests
scikit-learn
scipy
numpy